+++ test/meetingreadinesschecker/DefaultMeetingReadinessChecker.test.ts
@@ -281,6 +281,16 @@ describe('DefaultMeetingReadinessChecker', () => {
       expect(audioOutputFeedback).to.equal(CheckAudioOutputFeedback.Succeeded);
     });
 
+    it('successful after playing tone with string deviceId', async () => {
+      const successCallback = (): Promise<boolean> => Promise.resolve(true);
+
+      const audioOutputFeedback: CheckAudioOutputFeedback = await meetingReadinessCheckerController.checkAudioOutput(
+        'deviceId',
+        successCallback
+      );
+      expect(audioOutputFeedback).to.equal(CheckAudioOutputFeedback.Succeeded);
+    });
+
     it('can be called multiple times', async () => {
       const successCallback = (): Promise<boolean> => {
         return new Promise(resolve => {
@@ -436,7 +446,17 @@ describe('DefaultMeetingReadinessChecker', () => {
     it('checks for 240 x 320 resolution - success', async () => {
       domMockBehavior.getUserMediaSucceeds = true;
       const cameraResolutionFeedback: CheckCameraResolutionFeedback = await meetingReadinessCheckerController.checkCameraResolution(
-        new MediaDeviceInfo(),
+        getMediaDeviceInfo('1', 'videoinput', 'label', 'group-id'),
+        240,
+        320
+      );
+      expect(cameraResolutionFeedback).to.equal(CheckCameraResolutionFeedback.Succeeded);
+    });
+
+    it('checks for 240 x 320 resolution with string input - success', async () => {
+      domMockBehavior.getUserMediaSucceeds = true;
+      const cameraResolutionFeedback: CheckCameraResolutionFeedback = await meetingReadinessCheckerController.checkCameraResolution(
+        'deviceId',
         240,
         320
       );
@@ -447,7 +467,7 @@ describe('DefaultMeetingReadinessChecker', () => {
       domMockBehavior.getUserMediaSucceeds = false;
       domMockBehavior.getUserMediaResult = UserMediaState.Failure;
       const cameraResolutionFeedback: CheckCameraResolutionFeedback = await meetingReadinessCheckerController.checkCameraResolution(
-        new MediaDeviceInfo(),
+        getMediaDeviceInfo('1', 'videoinput', 'label', 'group-id'),
         240,
         320
       );
@@ -458,7 +478,7 @@ describe('DefaultMeetingReadinessChecker', () => {
       domMockBehavior.getUserMediaSucceeds = false;
       domMockBehavior.getUserMediaResult = UserMediaState.OverconstrainedError;
       const cameraResolutionFeedback: CheckCameraResolutionFeedback = await meetingReadinessCheckerController.checkCameraResolution(
-        new MediaDeviceInfo(),
+        getMediaDeviceInfo('1', 'videoinput', 'label', 'group-id'),
         7680,
         4320
       );
@@ -471,7 +491,7 @@ describe('DefaultMeetingReadinessChecker', () => {
       domMockBehavior.getUserMediaSucceeds = false;
       domMockBehavior.getUserMediaResult = UserMediaState.NotAllowedError;
       const cameraResolutionFeedback: CheckCameraResolutionFeedback = await meetingReadinessCheckerController.checkCameraResolution(
-        new MediaDeviceInfo(),
+        getMediaDeviceInfo('1', 'videoinput', 'label', 'group-id'),
         240,
         320
       );
@@ -487,7 +507,7 @@ describe('DefaultMeetingReadinessChecker', () => {
         meetingSession
       );
       const cameraResolutionFeedback: CheckCameraResolutionFeedback = await meetingReadinessCheckerController.checkCameraResolution(
-        new MediaDeviceInfo(),
+        getMediaDeviceInfo('1', 'videoinput', 'label', 'group-id'),
         240,
         320
       );